requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
gradio>=4.0.0